Move UserCreate back into Components folder
